# catalog/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Catalog API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/{category}/sub_category/create": {
            "post": {
                "tags": [
                    "SubCategories"
                ],
                "summary": "Создание новой подкатегории",
                "operationId": "c2e7049b2b5118d666260f1ead263aff",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "ID категории",
                            "type": "integer",
                            "example": "3"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Игрушки"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Массив подкатегории с данными",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SubCategory"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/admin/{category}/{sub_category}": {
            "put": {
                "tags": [
                    "SubCategories"
                ],
                "summary": "Обновление подкатегории",
                "operationId": "77fe49fd5b46bf9e7e0ce26d1f61913a",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "ID категории",
                            "type": "integer",
                            "example": "3"
                        }
                    },
                    {
                        "name": "sub_category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "ID подкатегорий",
                            "type": "integer",
                            "example": "3"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Игрушки"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Массив подкатегории с товарами",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SubCategoryResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "SubCategories"
                ],
                "summary": "Удаление подкатегории",
                "operationId": "0ba613e62259fa1e395f9aa3db3307db",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "ID категории",
                            "type": "integer",
                            "example": "3"
                        }
                    },
                    {
                        "name": "sub_category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "ID подкатегорий",
                            "type": "integer",
                            "example": "3"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Подкатегория удалена"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Category": {
                "description": "Категория для продуктов",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "ID категории",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Название категории",
                        "type": "string",
                        "example": "Футболка"
                    }
                },
                "type": "object"
            },
...
            "Product": {
                "description": "Класс продукта",
                "required": [
                    "id",
                    "name",
                    "description",
                    "price",
                    "category_id",
                    "sub_category_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID продукта",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Название продукта",
                        "type": "string",
                        "example": "Футболка"
                    },
                    "description": {
                        "description": "Описание продукта",
                        "type": "text",
                        "example": "Хлопок, сертифицированный по экологическим стандартам, при его производстве не используются агрессивные химикаты. Выбирая такой товар, вы проявляете заботу о себе и о планете."
                    },
                    "price": {
                        "description": "Цена продукта",
                        "type": "integer",
                        "example": "9 990"
                    },
                    "previousPrice": {
                        "description": "Прежняя цена продукта",
                        "type": "integer",
                        "example": "10 990"
                    },
                    "sub_category_id": {
                        "description": "ID подктаегории товара",
                        "type": "integer",
                        "example": "3"
                    }
                },
                "type": "object"
            },
            "SubCategory": {
                "description": "Подкатегория для продуктов",
                "required": [
                    "id",
                    "name",
                    "category_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID подкатегории",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Название подкатегории",
                        "type": "string",
                        "example": "Футболка"
                    },
                    "category_id": {
                        "description": "Основная категория для подкатегории",
                        "type": "integer",
                        "example": "3"
                    },
                    "products": {
                        "description": "ID подкатегории товара",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Product"
                        },
                        "example": [
                            {
                                "id": 1,
                                "name": "Худи",
                                "description": "Ea ut consequatur natus sed occaecati hic. Adipisci repellendus id cum itaque quis aliquid. Aut sunt suscipit in sequi tenetur. Sit illo quo velit quia quia possimus.",
                                "price": "KZT 34,227.00",
                                "previousPrice": "KZT 92,848.00",
                                "sub_category": "Верхняя одежда",
                                "category_id": "Одежда"
                            },
                            {
                                "id": 2,
                                "name": "Худи",
                                "description": "Suscipit et modi magnam in dolorum. Sit deserunt sunt aut non. Nulla nihil veritatis aut fuga consectetur quidem quae. Numquam earum excepturi sapiente et illo nisi sint.",
                                "price": "KZT 64,807.00",
                                "previousPrice": "KZT 42,079.00",
                                "sub_category": "Верхняя одежда",
                                "category_id": "Одежда"
                            }
                        ]
                    }
                },
                "type": "object"
            },
...
            "SubCategoryResource": {
                "description": "Подкатегория для продуктов",
                "required": [
                    "id",
                    "name",
                    "category_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID подкатегории",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Название подкатегории",
                        "type": "string",
                        "example": "Футболка"
                    },
                    "category_id": {
                        "description": "Основная категория для подкатегории",
                        "type": "integer",
                        "example": "3"
                    },
                    "products": {
                        "description": "Товары подктаегории",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        },
                        "example": [
                            {
                                "id": 1,
                                "name": "Худи",
                                "description": "Ea ut consequatur natus sed occaecati hic. Adipisci repellendus id cum itaque quis aliquid. Aut sunt suscipit in sequi tenetur. Sit illo quo velit quia quia possimus.",
                                "price": "KZT 34,227.00",
                                "previousPrice": "KZT 92,848.00",
                                "sub_category": "Верхняя одежда",
                                "category_id": "Одежда"
                            },
                            {
                                "id": 2,
                                "name": "Худи",
                                "description": "Suscipit et modi magnam in dolorum. Sit deserunt sunt aut non. Nulla nihil veritatis aut fuga consectetur quidem quae. Numquam earum excepturi sapiente et illo nisi sint.",
                                "price": "KZT 64,807.00",
                                "previousPrice": "KZT 42,079.00",
                                "sub_category": "Верхняя одежда",
                                "category_id": "Одежда"
                            }
                        ]
                    }
                },
                "type": "object"
            }
        }
    },
...
}